% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blizz.R
\name{blizz}
\alias{blizz}
\title{Extract Data from Blizzards API}
\usage{
blizz(endpoint, region = "us", json = FALSE)
}
\arguments{
\item{endpoint}{An endpoint provided by Blizzards API documentation}

\item{region}{API data is limited to specific regions. For example, US APIs
accessed through us.battle.net only contain data from US battlegroups and
realms. See the dataset \code{regions} for more info.}

\item{json}{Logical TRUE/FALSE to return a JSON object or a R object,
defaults to R.}
}
\description{
Main function for interactive with Blizzards API. At the bare
minimum, you must provide the endpoint (see: Blizzards API docs). You may
also suggest a valid region and locale. By default, \code{blizz} defaults to
the US. Finally, you may view the data as a JSON object i.e.
\code{json = TRUE} or as a R object (default) i.e. \code{json = FALSE}.
}
