% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blizz_creds.R
\name{blizz_creds}
\alias{blizz_creds}
\title{Install All Required Blizzard Credentials in Your \code{.Renviron} File for
Repeated Use}
\usage{
blizz_creds(id, secret, overwrite = FALSE, install = TRUE,
  refresh = FALSE)
}
\arguments{
\item{id}{The CLIENT ID provided to you by Blizzard formated in quotes. A client id can be acquired at \url{https://develop.battle.net/}}

\item{secret}{The CLIENT SECRET provided to you by Blizzard formated in quotes. A client id can be acquired at \url{https://develop.battle.net/}}

\item{overwrite}{If this is set to TRUE, it will overwrite an existing Blizzard credentials that you already have in your \code{.Renviron} file.}

\item{install}{if TRUE, will install the key in your \code{.Renviron} file for use in future sessions.  Defaults to FALSE.}

\item{refresh}{If this is set to TRUE, the CLIENT_AUTH_TOKEN will be refreshed. Blizzard Auth Tokens expire after 24 hours.}
}
\description{
This function will add your BLIZZARD CLIENT ID,
BLIZZARD SECRET ID and BLIZZARD AUTH TOKEN to your \code{.Renviron} file so
it can be called securely without being stored in your code. After you have
installed your credentials, it can be called any time by typing
\code{Sys.getenv("BLIZZARD_CLIENT_ID")} and can be used in package functions
by simply typing BLIZZARD_CLIENT_ID. If you do not have an \code{.Renviron}
file, the function will create on for you. If you already have an
\code{.Renviron} file, the function will append the key to your existing
file, while making a backup of your original file for disaster recovery purposes.
}
