% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d3acts}
\alias{d3acts}
\title{Diablo 3 Acts}
\format{
A data frame with 39 rows and 6 variables:
\describe{
  \item{id}{Act ID: \emph{87700, 72095, ...}}
  \item{act}{Act: \emph{Act I, Act II, ...}}
  \item{act_slug}{Act Slug: \emph{act-ii, act-ii, ...}}
  \item{number}{Act Number: \emph{1, 2, ...}}
  \item{name}{Act Name: \emph{The Fallen Star, The Legacy of Cain, ...}}
  \item{name_slug}{Act Name Slug: \emph{the-fallen-star, the-legacy-of-cain, ...}}
}
}
\source{
\url{https://develop.battle.net/documentation/api-reference}
}
\usage{
d3acts
}
\description{
A dataset containing all available Acts in the Diablo 3 storyline.
}
\keyword{datasets}
