% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blizz.R
\name{blizz}
\alias{blizz}
\title{Extract Blizzard Game Data}
\usage{
blizz(
  endpoint,
  locale = "en_US",
  namespace = NULL,
  json = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{endpoint}{An endpoint provided by Blizzards API documentation}

\item{locale}{All available API resources provide localized strings using the
locale query string parameter. Supported locales vary from region to region
and align with those supported on Blizzard community sites. Defaults to
\code{"en_US"}.}

\item{namespace}{Namespaces in Game Data and Profile APIs that allow JSON
documents to be published contextually in relation to a specific patch or
point in time. Defaults to \code{NULL}.}

\item{json}{Logical TRUE/FALSE to return a JSON object or a R object,
defaults to R.}

\item{quiet}{If \code{FALSE}, request information will not print. Defaults to
\code{TRUE}.}
}
\description{
Main function to grab data from Blizzard's API. At the bare
minimum, you must provide the endpoint (see: Blizzard's API docs). You may
also provide a valid locale and namespace. Finally, you may view the data as
a JSON object i.e. \code{blizz("/some/endpoint", json = TRUE)} or as a R
object (default) i.e. \code{blizz("/some/endpoint")}.
}
